[#79] Added `stalker.db.setup` module and moved all the DB setup routines to it to remove all the contents from the `__init__.py` which was suboptimal as the contents were getting importat upon importing any submodule under `stalker.db`.
